server/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
groq==1.0.0
psycopg2-binary==2.9.11
PyPDF2==3.0.1
sentence-transformers==5.1.2
numpy>=2.1.2
python-dotenv==1.2.1
